fix: adjust exception arguments
