+++ src/pipes/zod/zod-validation.exception.ts
@@ -1,0 +1,16 @@
+import { BadRequestException, HttpStatus } from '@nestjs/common';
+import { ZodError } from 'zod';
+
+export class ZodValidationException extends BadRequestException {
+  constructor(private error: ZodError) {
+    super({
+      statusCode: HttpStatus.BAD_REQUEST,
+      message: 'Validation failed',
+      cause: error.issues,
+    });
+  }
+}
+
+export const createZodValidationException = (error: ZodError): Error => {
+  return new ZodValidationException(error);
+};
