+++ src/pipes/zod/zod-validation.pipe.ts
@@ -1,13 +1,7 @@
 import { PipeTransform, Injectable } from '@nestjs/common';
-import { Schema, SafeParseError } from 'zod';
-import { HttpStatus } from '@nestjs/common';
+import { Schema } from 'zod';
 
-export const createZodException = <T>({ error }: SafeParseError<T>) => {
-  return {
-    statusCode: HttpStatus.BAD_REQUEST,
-    message: error.issues,
-  };
-};
+import { createZodValidationException } from './zod-validation.exception';
 
 @Injectable()
 export class ZodValidationPipe implements PipeTransform {
@@ -17,7 +11,7 @@ export class ZodValidationPipe implements PipeTransform {
     const result = this.schema.safeParse(value);
 
     if (!result.success) {
-      throw createZodException(result);
+      throw createZodValidationException(result.error);
     }
 
     return result.data;
